updated version to 2.0.1-SNAPSHOT changed package to ours
