+++ src/main/groovy/org/agileworks/elasticsearch/river/csv/CSVRiver.java
@@ -1,4 +1,4 @@
-package org.elasticsearch.river.csv;
+package org.agileworks.elasticsearch.river.csv;
 
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.bulk.BulkRequestBuilder;
