+++ src/main/groovy/org/agileworks/elasticsearch/river/csv/ConfigurationException.groovy
@@ -1,4 +1,4 @@
-package org.elasticsearch.river.csv
+package org.agileworks.elasticsearch.river.csv
 
 import groovy.transform.CompileStatic
 import groovy.transform.InheritConstructors
