+++ src/main/resources/es-plugin.properties
@@ -1,2 +1,2 @@
-plugin=org.elasticsearch.plugin.river.csv.CSVRiverPlugin
+plugin=org.agileworks.elasticsearch.plugin.river.csv.CSVRiverPlugin
