+++ src/test/groovy/org/agileworks/elasticsearch/river/csv/CSVConnectorTest.groovy
@@ -1,4 +1,4 @@
-package org.elasticsearch.river.csv
+package org.agileworks.elasticsearch.river.csv
 
 import org.elasticsearch.river.RiverSettings
 import spock.lang.Specification
