+++ src/test/groovy/org/agileworks/elasticsearch/river/csv/ConfigurationTest.groovy
@@ -1,5 +1,7 @@
-package org.elasticsearch.river.csv
+package org.agileworks.elasticsearch.river.csv
 
+import org.agileworks.elasticsearch.river.csv.Configuration
+import org.agileworks.elasticsearch.river.csv.ConfigurationException
 import org.elasticsearch.river.RiverSettings
 import spock.lang.Specification
